requirements.txt
Flask==3.0.0
python-dotenv==1.2.1
Flask-SQLAlchemy==3.1.1
pytest==8.4.2
pytest-cov==7.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
sentry-sdk[flask]==2.19.2
marshmallow==3.23.2
flask-limiter==3.8.0
Flask-Login==0.6.3
bcrypt==4.1.2
Flask-Migrate==4.0.5
openai==1.59.9
